+++ lib/src/features/rest_timer/application/rest_timer_application.dart
@@ -1,0 +1,1 @@
+export 'rest_timer_controller.dart';